+++ deps/ar/ar.cpp
@@ -31,14 +31,15 @@ bool ar::Dictionaries::check(std::string word) {
 
 void ar::Dictionaries::add(Dictionary d) { this->dicts.push_back(d); }
 
-LongForm expand_known_abbr(NonDictWords &nonDictWords, ar::Dictionaries D, NonDictWords *toExpand) {
+LongForm expand_known_abbr(NonDictWords &nonDictWords, ar::Dictionaries D,
+                           NonDictWords *toExpand) {
   LongForm retvec;
 
   for (auto it = 0; it < nonDictWords.size(); it++) {
     auto each = nonDictWords[it];
     try {
       auto word = D.known_abbr[each];
-      if(word.size() == 0) {
+      if (word.size() == 0) {
         toExpand->push_back(each);
         continue;
       }
@@ -94,7 +95,7 @@ LongForm ar::expansion_matching(NonDictWords nonDictWords, Dictionaries D) {
     return retvec;
   }
 
-  for(auto word: toExpand) {
+  for (auto word : toExpand) {
     INFO("HAVE TO EXPAND:" << word)
   }
 
@@ -107,31 +108,44 @@ LongForm ar::expansion_matching(NonDictWords nonDictWords, Dictionaries D) {
       int consonants = 0;
       check_vowels_consonants(token, &vowels, &consonants);
       if (vowels >= consonants) {
-        phi = [&token, &D](string &tok, string &word) { 
-          if(tok.length() != token.length()) {
+        // if the token is a prefix
+        phi = [&token, &D](string &tok, string &word) {
+          if (tok.length() != token.length()) {
             return -1;
           }
-          for(auto i = 0; i < tok.length(); i++) {
-            if(word[i] != tok[i]) {
+          for (auto i = 0; i < tok.length(); i++) {
+            if (word[i] != tok[i]) {
               return -1;
             }
           }
-          return 2;
+          return 1;
         };
       } else {
         phi = [&token, &D](string &tok, string &word) {
-          int counter = 0;
           int pos = 0;
-          if(tok.length() != token.length()) {
-            return -1;
+          if (tok[pos] != word[pos]) {
+            auto n_pos = token.find(tok);
+            if(n_pos != string::npos) {
+              pos = n_pos;
+            } else {
+              return -1;
+            }
           }
-          for(counter = 0; counter < tok.length(); counter++) {
-            for(auto j = pos; j < word.length(); j++) {
-              if(tok[counter] == word[j]) {
+          std::cout << tok << " " << word << std::endl;
+          int top_counter = 0;
+          int counter = 0;
+          // Ensure that the word contains all the needed letters
+          for (counter = 0; counter < token.length(); counter++) {
+            for (auto j = pos; j < word.length(); j++) {
+              if (token[counter] == word[j]) {
                 pos = j;
+                top_counter++;
                 break;
               }
             }
+          }
+          if (top_counter != counter) {
+            return -1;
           }
           return 1;
         };
@@ -248,7 +262,6 @@ Lmatch ar::split_matching(string ident, Dictionaries D, Lmatch *matches) {
       INFO("SPLIT_MATCHING DONE")
       break;
     }
-
   }
   Lmatch retvec;
   for (auto each : *matches) {
@@ -267,4 +280,128 @@ Lmatch ar::split_matching(string ident, Dictionaries D, Lmatch *matches) {
                      [](const std::string &str) { return str.size() == 0; }),
       retvec.end());
   return retvec;
+}
+
+#include <filesystem>
+#include <fstream>
+#include <math.h>
+#include <ostream>
+
+std::unordered_map<std::string, std::string> get_abbr(string file) {
+  std::unordered_map<std::string, std::string> retvec;
+
+  std::fstream ifile;
+  ifile.open(file);
+
+  if (!ifile.is_open()) {
+    std::cout << "Could not open file " << file << std::endl;
+    exit(1);
+  }
+
+  string line;
+  string delim = ";";
+  while (std::getline(ifile, line)) {
+    int first = line.find(delim);
+    string word = line.substr(0, first);
+    line.erase(0, first + delim.length());
+    string other = line;
+    retvec.insert({word, other});
+  }
+
+  return retvec;
+}
+
+ar::Dictionary get_dict(string file) {
+  ar::Dictionary retvec;
+
+  std::fstream ifile;
+  ifile.open(file);
+
+  if (!ifile.is_open()) {
+    std::cout << "Could not open file " << file << std::endl;
+    exit(1);
+  }
+  int file_size = std::count(std::istreambuf_iterator<char>(ifile),
+                             std::istreambuf_iterator<char>(), '\n');
+
+  ifile.clear();
+  ifile.seekg(0);
+
+  string line;
+  string delim = ";";
+  while (std::getline(ifile, line)) {
+    int first = line.find(delim);
+    string word = line.substr(0, first);
+    line.erase(0, first + delim.length());
+    int other = -1;
+    try {
+      other = std::stoi(line.substr(0, first));
+    } catch (std::invalid_argument e) {
+      other = (word.length() / std::log(word.length()));
+    }
+    retvec.insert({word, other});
+  }
+
+  return retvec;
+}
+
+ar::Dictionaries get_true() {
+  ar::Dictionaries D;
+  auto eng = get_dict("test.txt");
+  auto it = get_dict("List_of_computing_and_IT_abbreviations");
+  D.add(eng);
+  D.add(it);
+  D.known_abbr = get_abbr("known_abbr.txt");
+  return D;
+}
+
+auto D = get_true();
+std::string ar::do_ar(std::string token) {
+  RESULT("STARTED: " << token)
+
+  Lmatch matches;
+  auto labels = ar::split_matching(token, D, &matches);
+
+  auto thing = ar::expansion_matching(labels, D);
+
+  // Replace the words in the identifier
+  for (auto each : matches) {
+    size_t index = 0;
+    while (true) {
+      /* Locate the substring to replace. */
+      index = token.find(each, index);
+      if (index == std::string::npos)
+        break;
+
+      /* Make the replacement. */
+      token.replace(index, each.size(), each + "_");
+
+      /* Advance index forward so the next iteration doesn't pick it up as
+       * well.
+       */
+      index += each.size();
+    }
+  }
+
+  for (auto &[each, word] : thing) {
+    size_t index = 0;
+    while (true) {
+      /* Locate the substring to replace. */
+      index = token.find(each, index);
+      if (index == std::string::npos)
+        break;
+
+      /* Make the replacement. */
+      token.replace(index, each.size(), word + "_");
+
+      /* Advance index forward so the next iteration doesn't pick it up as
+       * well.
+       */
+      index += word.size();
+    }
+  }
+
+  RESULT(token)
+
+  return token;
 }
