+++ deps/ar/ar.hpp
@@ -43,6 +43,8 @@ struct Dictionaries {
     return dicts.end(); } // 200 is out of bounds
 };
 
+std::string do_ar(std::string word);
+
 LongForm expansion_matching(NonDictWords nonDictWords, ar::Dictionaries D);
 
 Lmatch string_matching(std::string token, ar::Dictionary D, Phi phi,
